# hw10 done, improve performance