--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -25,13 +25,11 @@ var json = jsoniter.ConfigCompatibleWithStandardLibrary
 func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
 	result := make(DomainStat)
 	scanner := bufio.NewScanner(r)
-	scanner.Split(bufio.ScanLines)
 	suffix := "." + domain
 	var user User
 	for scanner.Scan() {
-		user.Email = ""
 		if err := json.Unmarshal(scanner.Bytes(), &user); err != nil {
-			continue
+			return nil, err
 		}
 		if strings.HasSuffix(user.Email, suffix) {
 			result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]++