@farir1408 2021-10-12
Здравствуйте!

Спасибо за выполнение домашнего задания.
Хотелось бы видеть новые тесты.

Отличная работа, принято 8/10.